Added rough REST style mappings.
It aint spring, but hey it's kinda like express?
Annotation support requires reflection, which isn't worth the trouble
currently. This resolves issue #7 and also raises another issue with the
current StatusCode enum, not easy way to check if code is within a
status code range, i.e. the code is a 4XX code. The current method of
taking the string version and slicing it is awful.